requirements.txt
numpy
matplotlib
pyserial
jupyterlab
notebook
ipympl
jupyter
